requirements.txt
flask>=2.1.2
boto3>=1.2.4
python-dotenv>=0.20.0